tdf#165047 sw mso-compat layout: always honour contextualSpacing
This improves 25.8 commit ae7900dd42a65aaf60df6b21b9ad511496b209d9
tdf#164095 sw: fix missing top margin on paragraph
               after changing page style
which was backported to 24.8.5.

The problem is that you can't just simply apply the top margin.
Even though it may be across a page break,
or even in a different section,
the fact that the previous paragraph is the same style
means that we simply don't add any upper spacing to this paragraph.

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf165047_contextualSpacingTopMargin

Change-Id: I12712e5421e20ea0757c59029c6890bff9d81b94
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183643
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183650
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>