Impress: implement DontTerminateEdit support
Name FN_PARAM1 in calc so it is clear what it does.
It is used as DontTerminateEdit indicator used eg. in
LOK case when we do saving in the background. In Impress
when we edit textbox and close presentation - sometimes it
is not saved because we didn't press Enter to commit the change
inside. In LOK case we want to save the content to not allow
data loss in unexpected connection close.

Change-Id: Ibcdcd4a0d52eb36e0b1004971dcb73f58c35459e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183652
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>