cool#11357 sw redline reinstate: add to the context menu for text
Right click on a tracked insert, you can accept and reject the change,
but not reinstate it.

The two extremes would be to not allow reinstate there at all and to add
all of reinstate, reinstate-and-next and reinstate-all to the context
menu.

Be consistent with accept/reject and only add
.uno:ReinstateTrackedChange to avoid cluttering the context menu too
much.

Change-Id: I8f05eedc066a67327e0c7ad25a0e6480867d872b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183664
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins