SAL_INFO when vnd.sun.star.expand is considered exotic
As discussed in the appertaining comments at
<https://gerrit.libreoffice.org/c/core/+/176645/3#message-1209b271ac4a330311eb90b2d6c16bbff57842ef>,
836d73a65180d89a077e36457f1f3aa1698c2058 "consider VndSunStarExpand an exotic
protocol" may have been a bit brittle:  "[vnd.sun.star.expand] originated from
the UNO configuration infrastructure, but indeed hard to tell if it then found
reasonable uses outside that area.  I guess the best we can do is get it in,
maybe mention it in the release notes, and see."  Turns out it is used at least
to handle %origin% in extensions, where that change caused
<https://bugs.documentfoundation.org/show_bug.cgi?id=165851> "Installing custom
default template via extension stopped working".

So to be better prepared to track down potential further breakage caused by that
change, note cases where it (legitimately or problematically) hits with a
SAL_INFO.

Change-Id: If742f4f0b0f223e5acf346d294567ccc09eaf49b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183659
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <stephan.bergmann@allotropia.de>