editeng a11y: Let para manager dispose its child
Instead of letting AccessibleParaManager dispose
itself in AccessibleEditableTextPara::SetEditSource when
called with a null edit source, let the AccessibleParaManager
that created the paragraph and can therefore be considered the
owner (see AccessibleParaManager::CreateChild) dispose it
again in AccessibleParaManager::Release.
This is consistent with how it also explicitly disposes
all children in AccessibleParaManager::Dispose.

This is the only place that calls AccessibleEditableTextPara::SetEditSource
with a null param, so this change shouldn't change behavior
in any way.

Change-Id: I9d6ec8db69a789c18fd424e680624744b4913798
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183679
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>