editeng a11y: Use OAccessibleComponentHelper for paragraph
Let editeng's AccessibleEditableTextPara derive from
OAccessibleComponentHelper to make use of the
logic already implemented there instead of having to
implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

The existing AccessibleEditableTextPara::getBounds
implementations provide the necessary logic to implement
OAccessibleComponentHelper::implGetBounds,
so rename those accordingly (and drop the extra mutex
guard, as the base class takes care of that).

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the base class.

No change in behavior intended or observed in a quick
test with text in Impress slides and Accerciser and Orca
when using the gtk3 and qt6 VCL plugins on Linux.

Change-Id: I5cb47eb9c6738f4bd7807cbeef260151437c0862
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183681
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins