tdf#165941 DOCX import: fix unhandled exc while removing dummy para in table
The bugdoc can't be opened in Writer but Word considers it valid.

This went wrong in commit 97f9af714ea1c46e498fa99f7ca34fc1708d38a6
(Related: tdf#125038 DOCX import: fix unexpected linebreak inside IF
condition, 2019-10-30), because now we may have a range that is crossing
a table of section boundary: this is a valid range for a field start/end
in Word, but not in Writer.

Fix the problem by handling exceptions from createTextCursorByRange() in
DomainMapper_Impl::RemoveDummyParaForTableInSection(): we already
assumed previously that the created cursor may be an empty reference,
just the exception case was unexpected.

The bigger problem is that Writer fields can't expand to multi-paragraph
content while Word can do that, but that was a problem before the above
commit, so is not addressed here.

Change-Id: I718b702e68da5bf90eb4af95bb179d2e33abadc3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183695
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>