tdf#147874 slow sheet switching with large chart2 (2)
ExplicitCategoriesProvider is rather expensive to create, and we do it
twice.

So store it on the BaseCoordinateSystem object and share it, which means
we only need to construct it once.

This reduces switching time from 2s to 1s.

Note that I am not sure that this is the best location to store this,
the entire architecture of chart2 is a little opaque.

Change-Id: Ic9dc9573f7495bcb76de14caa7cc7d5ad61d9ed1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183700
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins