tdf#130857 qt weld: Use QDoubleSpinBox logic when no parser set
    commit 210c69cf7aaa314f932a9debe0e64ba2789f2365
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 20 19:22:52 2025 +0100

        tdf#130857 qt weld: Implement spin button input validation

implemented input validation for QtInstanceSpinButton
and QtInstanceFormattedSpinButton.

If a function to parse the text is set,
QtDoubleSpinBox::validate uses that to validate
the input, as the commit message mentions:

> If a handler for parsing the text is set and
> it can successfully parse the text, return
> QValidator::Acceptable, otherwise
> QValidator::Intermediate (allowing to continue
> editing/typing).

This implementation however assumes that the parser
function will return a std::optional containing
a value when the input is valid.

That assumption is correct for QtInstanceFormattedSpinButton.
For QtInstanceSpinButton however, there's also the
case where no formatter has been set, in which case
weld::SpinButton::parse_text returns TRISTATE_INDET
(s. its documentation).

So far, this resulted in QtInstanceSpinButton::convertTextToDouble
- which was set for QtDoubleSpinBox::m_aParseTextFunction, i.e. as
the function to parse the text - to always return an empty std::optional.

As a consequence, the value wouldn't be accepted
by QtDoubleSpinBox, e.g. for the following scenario
since the above-mentioned commit when using the qt5 or qt6 VCL
plugin with SAL_VCL_QT_USE_WELDED_WIDGETS=1:

1) "Tools" -> "Options" -> "Advanced" -> "Open Expert Configuration"
2) search for "LibreOfficeTheme"
3) double-click on the config entry (of type short)
4) in the spinbox, manually type a value of 2
5) press Tab to move focus out of the spinbox

Result: The manually typed value of 2 gets reset to
the previous value (currently 0 by default) instead of
taking over the new value.

Fix this by extending QtDoubleSpinBox::m_aParseTextFunction
to use the same semantics as weld::SpinButton::parse_text,
i.e. return a Tristate value which allows to distinguish
between the 3 different outcomes (parser set and value
could be parsed, parser set and value could not be parsed,
no parser set) and in case it returns TRISTATE_INDET
(indicating that no parser is set), fall back to
the logic in the QDoubleSpinBox base class to try to
parse/validate the current text.

Change-Id: Ic811dae10ecfd624e8f2e26022a61a95365152c2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183713
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>