tdf#166040 package: ZipFile::recover() find DD without PK78 signature
The PK78 signature is optional, and if the sizes in the local file
header are 0 then without the DD the recovery won't do anything useful.

If the latest local file header has the DD bit set in flags, then try to
find the DD before the next thing that has a PK?? signature.

Change-Id: I149a87805f7bcb6216fc909c20ece600ce77956d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183722
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins