tdf#165492 sw mso layout: add bottom border padding to cell minHeight
at least for MS Word formats.

This improves 25.2 commit 4f5b896d004af023f210f235aeafa9abc50068d2
    tdf#155229 Calculate row height incl. border if 'atLeast' is set
where it seemed to work if (as is typical)
the top and bottom padding were identical.

However, the top padding actually has no effect
on the minimal cell size - only on the actual fitted size.
Yet, the bottom padding DOES affect the minimal cell size.

[Bottom padding should also be added on top of the "exact" cell height,
but I didn't see a place where both of these could be handled
with a single chunk of code.]

make CppunitTest_sw_ooxmlexport22 \
   CPPUNIT_TEST_NAME=testTdf165492_atLeastWithBottomSpacing

A follow-up commit (for tdf#165424) is needed to fix this unit test,
where we don't import the overriding row definitions
that would remove the bottom border padding for all-but-the-first row.
    TestTdf164907_rowHeightAtLeast

Change-Id: Ib987defc6187171de05884cb0ced3273cdc57b9f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183737
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins