editeng a11y: Use OAccessibleComponentHelper for bullet
Let editeng's AccessibleImageBullet derive from
OAccessibleComponentHelper to make use of the
logic already implemented there instead of having to
implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

The existing AccessibleImageBullet::implGetBounds
implementations provide the necessary logic to implement
OAccessibleComponentHelper::implGetBounds, so mark it
as overriding the base class method.

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the base class.

No change in behavior intended or observed in a quick
test with Impress and Accerciser and the gtk3 VCL plugin
on Linux.

Sample steps to add a bullet item that shows
up in the a11y tree:

1) start Impress
2) Create a new slide (Ctrl+M)
3) type text "Hello world" for the first list item
4) right-click on paragraph to open context menu, select "Bullets and
   Numbering"
5) For "Type", select "Graphics"
6) select "Select image..." -> "Gallery" -> "bluearrow"
7) close dialog with "OK" button

Then, start Accerciser, and locate the "Image bullet"
object within LibreOffice's a11y tree.

Change-Id: Iba9d44288be96838e39d7cee2333f7a074deae8e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183783
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>