tdf#130857 qt weld: Support PasswordDialog
This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This dialog gets shown when opening a password-protected
file, allowing to type the password in order to open the
file.

When typing an incorrect password and confirming using
the OK button, the dialog is oddly shown as "Set Password"
dialog on top of the warning dialog ("The password is incorrect.
The file cannot be opened.") when using qt6, which is not the
case for qt5.
This will be addressed in a separate commit.

Change-Id: I9d1ffeaf48380aa4849e3cb771f6ff517d588790
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183795
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins