Related: cool#11226 fix 'record track changes' export to FODT
Create a Writer document, enable tracked changes, save to FODT, reopen,
changes are no longer tracked.

This seems to be an export problem, <text:tracked-changes
text:track-changes="true"> is simply not in the export result.

What happens is that the export wants to show insertions and hide
deletions during the export, and this happens in
SwXMLExport::exportDoc() for FODT while it happens in
SwXMLWriter::Write_() for ODT. This went wrong in commit
6af264883910fe31433b4164b1956f4f9ed75ecb (#94187# switch redline mode
outside of XML export components (or inside for each of them),
2001-11-08), where the mask to clear out the "show" bits was wrong for
the flat case and good for the package case.

Now that we actually set redline flags to hide deletions during FODT
export (or anything based on that, like the XHTML export), an additional
problem is that SetRedlineFlags() asserts in UnHide() for
CppunitTest_sw_xhtmlexport's testTdf66305, because there is no
StartAllAction() / EndAllAction() around it, while that is happening in
SwWriter::Write() for the ODT case. Fix this, too.

Change-Id: I227a883ecc5d68611d8aee5c80d65485b5f3c4c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183834
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>