sw: docx: add annotationRef on comments
Right now we never add <w:annotationRef/> to any comment when saving to
docx. According to the standard, not having it is the same as having it
in the beginning. But when round-tripping a document with multiple
comments without text from MS Word, the comments get mangled if the
annotationRef tags aren't kept. So it's better to always write the
implicit annotationRef at the beginning of comments.

Change-Id: I6c395bbc691a29533e2be9a93663993e083db65b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183856
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>