related tdf#166068: sd: make paste-special html paste into current textedit
fix HTML paste-special option creating a new text object even if
paste-special was used in a text-edit context. insert the text into the
current text edit context instead of creating a seperate object.

this is the current behavior for RTF & Flat XML Format paste options- so
it is was a bug for html paste option.

Change-Id: Ief9a19997a619b9f0b9c7855561de3f00de2df68
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183913
Tested-by: Jenkins
Reviewed-by: Sarper Akdemir <sarper.akdemir@allotropia.de>