sc a11y: Deduplicate code for AccessibleEventObject creation
Change ScAccessibleContextBase::CommitChange
from taking a `const AccessibleEventObject&`
param to taking the parameters needed to
create an AccessibleEventObject instead and
create that event in ScAccessibleContextBase::CommitChange
instead of having all callers do that themselves.

This also aligns this with
OAccessibleComponentHelper::NotifyAccessibleEvent
and therefore prepares for turning
ScAccessibleContextBase into an OAccessibleComponentHelper
subclass in an upcoming commit, to further deduplicate
XAccessibleEventBroadcaster and XAccessibleComponent
implementations.

For most cases, the event being created should be
unchanged, with few exceptions for cases that looked a
bit suspicious:

1) In ScAccessibleSpreadsheet::NotifyRefMode, a single
event object was created and then reused multiple times,
with just adjusting some of the members. While that
is fine in general, the AccessibleEventObject::OldValue
being set to the old/previous formula cell doesn't make
sense for the AccessibleEventId::SELECTION_CHANGED_WITHIN
and AccessibleEventId::SELECTION_CHANGED_ADD events, so
it is no longer set there. (For other events, the
doc in offapi/com/sun/star/accessibility/AccessibleEventId.idl
also doesn't always document an `OldValue` to be set
(or even says it should be empty), but at a quick glance,
winaccessibility might still process it if set, so leave
it in place for now.
(This should potentially be revisited independently, not
as part of a change like this that intends to not change
behavior for end users.)

2) In ScAccessibleSpreadsheet::Notify, pass an
empty/default-constructed uno::Any for the OldValue
of the AccessibleEventId::ACTIVE_DESCENDANT_CHANGED_NOFOCUS
event instead of an emtpy reference. The only handler in
AccDescendantManagerEventListener::HandleChildChangedNoFocusEvent
handles both the same way.
Similar for any events further down where the OldValue
would still be set as it's not explicitly unset.

No change in user-observable behavior was noticed in a quick test
of Calc with Orca and Accerciser and the gtk3 and qt6 VCL plugins
on Linux with this commit in place.

Change-Id: Ida3264b0935f544d4ce787b71684af148e47f86b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183931
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>