sc a11y: Use OAccessibleComponentHelper
Let ScAccessibleContextBase derive from
OAccessibleComponentHelper to make use of the
logic already implemented there instead of having to
implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

The existing ScAccessibleContextBase::GetBoundingBox provides
necessary logic to implement OAccessibleComponentHelper::implGetBounds,
so simply call it.

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the base class.

No change in behavior intended or observed in a quick
test with Calc using Orca and Accerciser with the gtk3
and qt6 VCL plugins on Linux.

Change-Id: I3d6db0e8c3b7c165dbbf9399e9b7c5b789cbfac5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183938
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>