tdf#166107: remove unnecessary resetPrintArea()
In the mentioned regression, resetPrintArea() was responsible for reverting the print area in Calc, which was set by the Print Dialog,
back to the one defined by the Page Styles.
It was implemented in the print-related part that is used by all modules but this is a Calc-specific issue.
Therefore, in this commit, resetPrintArea() was removed and it made
more sense to fix the print area issue in the Calc modul.

from
commit 364f0bb1cac0e12f5f926857f61c2f329a353ec7
Author: Tibor Nagy <tibor.nagy.extern@allotropia.de>
Date:   Wed Feb 28 11:13:10 2024 +0100

    tdf#155218 sc: fix regression page orientation in print dialog

Change-Id: I3767c0685addefb9b2fb16bfc7cf967a0f332272
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183948
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
(cherry picked from commit eccf4cd57a1970b296e12296bf0f5a0f52e917a7)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184211
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>