tdf#55425, tdf#66580: sdext,pdfimport: Start detecting newer hybrids
There are two mechanisms in Hybrid PDFs to mark the data, one
uses a non-standardised entry in the PDF trailer, another newer one , uses
a well-known name for a standard PDF embedded file.

The import filter currently just detects the non-standard trailer
addition, not the embedded file case, however the export code
doesn't write the old trailer for some formats, e.g. PDF/UA
and there's also a desire to get rid of the non-standard stuff
(see tdf#66580 ).

Add a detection mechanism for the newer embeddedFile form,
keep the old trailer detect to be able to load ancient files
that were generated before the embeddedFile format was added.

This is the first step of opening the file and wiring up
PDFium.

Change-Id: I44c0ffeac5e62dc1345c604bd695a2c2da37260a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183765
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins