tdf#55425, tdf#66580: sdext,pdfimport: Filename/MIME lookup
Look up the filename of the embedded file and find the MIME type we
associate with it.  If the name doesn't match, then it's not a LO
hybrid file.
When PDFium can read the MIME type, we'll be able to check
this MIME type against what is in the PDF.

Change-Id: Iebccf67c1d86e7c6c16c2c1c9ed8d1c08df758fb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183951
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>