tdf#130857 qt weld: Support "GtkRadioMenuItem"
Implement support for GtkRadioMenuItem [1] menu items
in .ui files.

Reuse the logic implemented for "GtkMenuItem". In addition,
support the "group" property which specifies what radio button
group the item belongs to.
For GTK, this is the ID of another GtkRadioMenuItem object.
For Qt, use a QActionGroup. Let it be owned by the QMenu
that  the menu items belong to, so it is deleted together with
the menu.

(The Qt concept of QActionGroup is basically the same as
QButtonGroup for radio buttons, see also
QtBuilder::setRadioButtonGroup.)

From a quick test, VclBuilder doesn't seem to evaluate the
"group" property, but seems to use some logic that simply
considers adjacent radio menu items as part of the same group.

This will be used by Math's "Format" -> "Spacing" dialog,
for which support will be declared in an upcoming commit.

[1] https://docs.gtk.org/gtk3/class.RadioMenuItem.html
[2] https://docs.gtk.org/gtk3/property.RadioMenuItem.group.html

Change-Id: Id0eece0eef65b8c52f77996c67662584f5665726
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184024
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>