tdf#130857 Define widgets in Math's "Spacing" dlg in order
Switch the order in which the widgets in the .ui file inside
the "grid2" are defined so that the order matches the visual appearance
(and "left-attach"/"top-attach" properties order), which makes sure
that tab focus order with the Qt-based VCL plugins is correct as well
when using native Qt widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where
focus order is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

The dialog can be triggered in Math via "Format" -> "Spacing".

Change-Id: I00c4e4bc93bc398f2f8752a9ff5d62ebe4caf8e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184027
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>