tdf#130857 qt weld: Make scroll area child resizable
As the QScrollArea doc [1] says:

> The child widget must be specified with setWidget().

This was implemented in

    commit 89b00d3e34c93b4ea0701a321a3cecf57aa0e378
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Dec 14 19:45:51 2024 +0100

        tdf#130857 qt weld: Drop scroll area parent if redundant

However, that commit had the side-effect that
the info text inside the scroll area in the
"View" -> "User Interface" dialog was no longer
completely displayed, but looked cropped.

This is because the label still had the size needed
to display only the initial "Information:" text
specified in the .ui file
(cui/uiconfig/ui/toolbarmodedialog.ui), not the
longer text set later.

Make the widget resizable by calling
QScrollArea::setWidgetResizable [2]
after creating the scroll area.

This makes the auto-resize of the label
work so that the full text is displayed
again as expected.

[1] https://doc.qt.io/qt-6/qscrollarea.html
[2] https://doc.qt.io/qt-6/qscrollarea.html#widgetResizable-prop

Change-Id: I2fada7b5e2037d6631006dcabd0ce08dfff27795
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184033
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>