tdf#130857 qt weld: Evaluate "inconsistent" prop for checkbuttons
For check boxes/buttons, evaluate the
GtkToggleButton::inconsistent [1] property when
processing the .ui file and set corresponding
QCheckBox properties when true.

This is used e.g. in cui/uiconfig/ui/fontfragment.ui
which is used in the "Font Features" dialog (which is
not yet supported by QtInstanceBuilder, but this
commit prepares for it).

[1] https://docs.gtk.org/gtk3/property.ToggleButton.inconsistent.html

Change-Id: If9a2aaee64998f0fe41c93ca42e4108ff46733d7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184038
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>