tdf#130857 qt weld: Handle special combobox insert pos -1
weld::ComboBox::append calls weld::ComboBox::insert
with a special insert position of -1, so handle
that in QtInstanceComboBox::insert.

Passing -1 as is to QComboBox::insertItem results in
the new entry to be inserted at position 0 instead,
i.e. as the first instead of as the last entry.

(Seen in a WIP branch adding support for the
"Font Features" dialog.)

Change-Id: Ia0c88449ed938d9ccb25bfefabd7f5aa742eed28
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184039
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>