tdf#130857 qt weld: Forward mouse click/release events
Set QtInstanceWidget as event filter for its QWidget
and in QtInstanceWidget::eventFilter, handle
(single and double) mouse click and mouse release
events by converting the QMouseEvent to a corresponding
vcl MouseEvent and calling weld::Widget::signal_mouse_press
and weld::Widget::signal_mouse_release.

In a WIP branch (also containing various other changes)
for the "File" -> "Templates" -> "Manage Templates" dialog,
this results in the correct template getting highlighted when
clicking on it and opened when double-clicking on it in
thumbnail mode.

Change-Id: I507deb1ed4aa2ae39c5f2b14f8bcb1fc1069ccf3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184048
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>