tdf#130857 qt weld: Show drawing area tooltips
This is similar to what

    commit dd9096736e2e89cc533bc04c77bf539bcc5df6bc
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 13 21:31:04 2025 +0100

        tdf#130857 qt weld: Show IconView item tooltips

implemented for weld::IconView, but now for weld::DrawingArea.

Note that `aHelpArea` is an in/out param for
weld::DrawingArea::signal_query_tooltip, see also
the gtk3 implementation in
GtkInstanceDrawingArea::signalQueryTooltip.

In a WIP branch (also containing various other changes)
for the "File" -> "Templates" -> "Manage Templates" dialog,
this results in the tooltip for the currently hovered
template to appear when hovering with the mouse and waiting
for a few seconds in thumbnail mode.

Change-Id: Ic829d1d1984aedddc0eaccba7538937defea57f3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184053
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>