tdf#130857 address book source dialog: No more do delayed initialization
Posting a user event and initialize things when that
event gets processed doesn't work with the qt6 VCL
plugin when using native widgets (for which support
will be added in an upcoming commit). There, the
user event only gets processed after the dialog gets
closed, and then results in a crash when support for
the dialog is declared in
QtInstanceBuilder::IsUIFileSupported.

This seems similar to the scenario in

    commit 29d6065d9a687f1d86f26f53587e108f8c951a4a
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Sep 16 15:21:01 2024 +0200

        sfx2: Synchronously set control state in file picker on all platforms

As was done there, do the work right away instead.
Should that result in any unresponsiveness for certain
scenarios, those will have to be examined more closely.

(At least with the Bibliography database used for the data
source, all seems fine.)

Change-Id: Ic2d35df49b683c27c2b505018374f421699f6711
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184079
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>