tdf#130857 qt weld: Add initial tool button support
Handle "GtkToolButton" objects in QtBuilder by creating
a QToolButton widget for them.
Evaluate the GtkToolButton::icon-name property [1].

This is used e.g. in the spelling dialog ("Tools" -> "Spelling"),
.ui file: cui/uiconfig/ui/spellingdialog.ui
(Dialog not declared as supported yet in QtInstanceBuilder and
there's much more to do before it can be.)

Support for more properties can be added when needed.

[1] https://docs.gtk.org/gtk3/property.ToolButton.icon-name.html

Change-Id: I07c406d6257912a64eeaa781edf245ad16c87862
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184084
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>