tdf#130857 qt weld: Signal when tool button is clicked
    Change-Id: I07c406d6257912a64eeaa781edf245ad16c87862
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Apr 12 00:32:01 2025 +0200

        tdf#130857 qt weld: Add initial tool button support

implemented support for "GtkToolButton" objects in .ui files
and inserts the created QToolButtons in the QToolBar parent
if there is one.

Build upon this and signal when one of the tool buttons
inside a toolbar is clicked by calling weld::Toolbar::signal_clicked.

Set the GtkBuilder ID as the QObject::objectName for
all created objects in QtBuilder::makeObject.

Connect to the QToolButton::clicked signal for all
tool buttons in the toolbar, and retrieve their ID
using the QObject::objectName to be able to notify what
is the ID of the clicked button.

In an experimental/WIP branch to support the spelling dialog
that contains more changes, this commit results in the
clipboard content to be pasted when clicking the corresponding
button in the spelling dialog toolbar, and clicking the special
characters button now results in the special character dialog
to open as expected, while nothing would happen previously.

Change-Id: I75b5136c85d4d2c44a312e55624c36a12e87dc7c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184085
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>