sw a11y: Fire focus event for scrolled in shape
Setting the AccessibleEventObject::Source to
a Reference of the AccessibleShape after
checking it has the focused state, but then
sending the event using
SwAccessibleContext::FireAccessibleEvent, i.e.
to the accessible listeners of the SwAccessibleContext,
not those of the shape, seems odd.

Instead, notify the listeners of the AccessibleShape
that has the FOCUSED state set.

Change-Id: Ic9e828c79d0fd99dcf50ba57f9374c427ae2d39a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184149
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>