sw a11y: Get rid of need to static_cast
Instead of letting the two SwAccessibleMap::GetContextImpl
call the corresponding SwAccessibleMap::GetContext
implementations that return an XAccessible and casting
that to SwAccessibleContext, rather do it the other way
around:

Move the existing logic (that already has an
rtl::Referenc<SwAccessibleContext> to
SwAccessibleMap::GetContextImpl and let
SwAccessibleMap::GetContext call the former.

This avoids the need static_cast.

Change-Id: I55cc3502adccb8fd6eece1dab467c8295a8137f3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184163
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins