Make the fontconfig Windows patch apply to fontconfig 2.16.1
Note that this patch is for now just a start of work that is paused
for now. The handling of non-ASCII filenames in the external libraries
that in the CODA case are now used also on Windows is likely very
broken, and needs to be carefully checked and modified. The new
<tools/UnixWrappers.h> thing, with additional work, will probably be
used.

Change-Id: I30cb09747a8de8716608669a75c872b73812b3f0