tdf#130857 Move weld::EntryTreeView dummy method impls to base class
Instead of having the same dummy implementations for weld::EntryTreeView
implementations for methods from the weld::ComboBox that only trigger
asserts in all subclasses, move those to the weld::EntryTreeView base
class.

As I understand it, these are most likely not actually used, but only
there because of weld::EntryTreeView being a weld::ComboBox subclass
(but GtkInstanceEntryTreeView and SalInstanceEntryTreeView not
subclassing GtkInstanceComboBox and SalInstanceComboBox, so they
cannot use the existing implementations from those).

Change-Id: Ibf88663fc83142300aa14b39a1e0805bd6f688cb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184181
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>