vcl: Introduce ImportOutput to return BitmapEx or Animation
Instead of returning Graphic, which is fully memory managed and
registers into Memorymanager, instead return ImportOutput which
returns the bitmap as BitmapEx and animation as Animation objects.

The reason is that we don't want to add into a memory managed
Graphic after import, as that might not be the intended use and
can just cause trashing of Graphic. For example this happens when
we want to import inside (Imp)Graphic itself, where we need to
create a new (Imp)Graphic temporarily for no reason when swapping
in, and then assign it to the existing (Imp)Graphic.

This converts PNG, JPG, GIF, XPM, XBM filters to use ImportOutput.

Change-Id: I862ef4125987436798314baaca95b6228f5d1a10
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184396
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins