tdf#166179 sw: fix inconsistent UI state vs command exec for the para marker
Go to the paragraph marker of the 2nd paragraph in the bugdoc, which is
red & strike-through. The toolbar suggests that strike-through is on,
but if you click on it, nothing happens.

The reason for this is that the DOCX bugdoc has these character
properties on the "paragraph marker", which can have its own formatting
in Word, but not really in Writer. Writer's native way to do something
similar is to just host character properties on the paragraph, which is
not possible in Word. This is not a new problem, commit
5ba30f588d6e41a13d68b1461345fca7a7ca61ac (tdf#64222 sw: better DOCX
import/export of paragraph marker formatting, 2019-09-17) added the
RES_PARATR_LIST_AUTOFMT paragraph property to store the formatting of
the paragraph marker and commit 6287845eab339a8eb3c660d648d012e9f1dcce30
(tdf#127606 DOCX lists: fix unchangeable formatting, 2020-04-17) already
ran into this "can't get rid of formatting" problem, where the fix was
to clear the RES_PARATR_LIST_AUTOFMT paragraph property.

Improves this a little, so:
1) If you go to the end of the paragraph and the UI hints that you have
   the relevant formatting enabled, then adjust command dispatch, so
   e.g. the "strike-through is enabled" -> press the toggle button ->
   "strike-through is not disabled" use-case works; and
2) Given that this paragraph marker formatting is currently duplicated
   in the doc model (stored in both RES_PARATR_LIST_AUTOFMT and also as
   an empty hint at the paragraph end), make sure that applying character
   formatting at paragraph end for these paragraphs doesn't result in an
   inconsistent model (we used to only update the hint, but not
   RES_PARATR_LIST_AUTOFMT).
   Do this a bit similar to how
   SwUnoCursorHelper::SetCursorPropertyValue() handles
   RES_PARATR_LIST_AUTOFMT during the DOCX import.

Note that lcl_InsAttr() is used with an item set, so applying a single
formatting via a toolbar button or using the char format dialog works.
Even undo/redo works, because SwRegHistory is still on the stack by the
time we call SwTextNode::SetAttr(), so the undo action will be an
SwUndoAttr, which has a SwHistorySetFormat, with both a SwHistoryHint
and an SwFormatAutoFormat set. This is now easier to see, when the
SwUndoAttr's SwHistory xml dump is part of the doc model xml dump.

Change-Id: Iaa50d1d2b3c6599c4fd617dfd85c9e93e070e972
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184197
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>